+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Prevent display name in notification header from shrinking too much on small screens.
 - Prevent overflow in poll options.
 - Prevent overflow in tables within a post.
+- Don't show number of minutes in payment status when it is negative.
 
 ## [4.13.0] - 2025-11-13